chore: move read commands to the commands.go
